+++ Sources/Public/Components/SelectionControl/Selectors/NatRadioButton/NatRadioButtonContainer.swift
@@ -33,6 +33,24 @@ class NatRadioButtonContainer: UIControl, NatSelector {
         }
     }
 
+    var isIndeterminate: Bool {
+        get {
+            radioButton.isIndeterminate
+        }
+        set {
+            radioButton.isIndeterminate = newValue
+        }
+    }
+
+    var labelComponent: String? {
+        get {
+            radioButton.labelComponent
+        }
+        set {
+            radioButton.labelComponent = newValue
+        }
+    }
+
     override init(frame: CGRect = .zero) {
         super.init(frame: frame)
         setup()
@@ -52,13 +70,13 @@ class NatRadioButtonContainer: UIControl, NatSelector {
         addSubview(radioButton)
 
         radioButton.onTouchesBegan = { [unowned self] touches in
-            self.addPulseLayerAnimated(at: radioButton.center, in: layer,
-                                       withColor: isSelected ? unselectedPulsableColor : selectedPulsableColor,
+            self.addPulseLayerAnimated(at: self.radioButton.center, in: self.layer,
+                                       withColor: self.isSelected ? self.unselectedPulsableColor : self.selectedPulsableColor,
                                        removeAfterAnimation: false)
         }
 
         radioButton.onTouchesEnded = { [unowned self] touches in
-            self.removePulseLayer(layer: layer)
+            self.removePulseLayer(layer: self.layer)
         }
 
         addConstraints()