+++ CHANGELOG.md
@@ -2,6 +2,51 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.6.0](https://github.com/natura-cosmeticos/natds-ios/compare/3.5.0...3.6.0) (2021-03-16)
+
+
+### Features
+
+* add list item component ([5d0f3e9](https://github.com/natura-cosmeticos/natds-ios/commit/5d0f3e9e7b58610aa288b17f991ac109659cc498))
+* update list item for sample app ([eb04538](https://github.com/natura-cosmeticos/natds-ios/commit/eb045381ad8b6ff205bb5aba2f0171e2cd11c5b9))
+
+
+### Bug Fixes
+
+* fix animation for selected cell ([517ea72](https://github.com/natura-cosmeticos/natds-ios/commit/517ea722d940a0bacb9b16588bd8cefc0ae0f511))
+* fix animation when cell is selected ([78b0d99](https://github.com/natura-cosmeticos/natds-ios/commit/78b0d99d7dab44686686d221a8f9e9a82dfddfdf))
+* fix cell selection in sample app ([60940bb](https://github.com/natura-cosmeticos/natds-ios/commit/60940bb1261538b2cc985c1769221cd63a5bb23d))
+* fix deprecated notice for checkbox ([92155c1](https://github.com/natura-cosmeticos/natds-ios/commit/92155c1b4a6e4711e6e4dc8110f22cc0965dc149))
+* fix onClick default state and animation func ([d472f5c](https://github.com/natura-cosmeticos/natds-ios/commit/d472f5c6cf83da686c58feef3b7ae1a73636e6d6))
+
+## [3.5.0](https://github.com/natura-cosmeticos/natds-ios/compare/3.4.0...3.5.0) (2021-03-16)
+
+
+### Features
+
+* WIP add attributes to textfield and to sample app ([b2b5aff](https://github.com/natura-cosmeticos/natds-ios/commit/b2b5aff382403be426c0350d0fadfd67d498f3be))
+* WIP new attributes to textField ([9fa56fd](https://github.com/natura-cosmeticos/natds-ios/commit/9fa56fd325320d2276230f9d0643668d1a38a598))
+
+
+### Bug Fixes
+
+* fix textfield's iconbutton size ([b0f370a](https://github.com/natura-cosmeticos/natds-ios/commit/b0f370ac1ffd9e5947e4861006435ea0d93f6af9))
+* remove paste function for number keyboard textfields ([7813e0c](https://github.com/natura-cosmeticos/natds-ios/commit/7813e0ccb872e103add146734aef0796b3c9f3d5))
+
+## [3.4.0](https://github.com/natura-cosmeticos/natds-ios/compare/3.3.5...3.4.0) (2021-03-12)
+
+
+### Features
+
+* 🎸 Adds indeterminate state and separes checkbox ([92f4c5e](https://github.com/natura-cosmeticos/natds-ios/commit/92f4c5e11dee04dab641c3bd30d3acd036b00907))
+
+### [3.3.5](https://github.com/natura-cosmeticos/natds-ios/compare/3.3.4...3.3.5) (2021-03-05)
+
+
+### Bug Fixes
+
+* 🐛 Fixes some implicit selfs  to run in xcode 11.5 ([131fdfc](https://github.com/natura-cosmeticos/natds-ios/commit/131fdfc81bb03560991a51795f80d0b34e69e5c1))
+
 ### [3.3.4](https://github.com/natura-cosmeticos/natds-ios/compare/3.3.3...3.3.4) (2021-02-25)